pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.1.0"
description = "gpt-like-a-dev"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.12",
    "jupyter>=1.1.1",
    "openai",
    "pydantic>=2.11.4",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.18",
]
